Move full map and object decoder into tools
